atlas-ares/requirements/base.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.32
pydantic==2.8.2
bcrypt==4.1.3
PyJWT==2.9.0
requests==2.32.3
python-dotenv==1.0.1
